add messageStoreHelper to refact messageStore
